+++ CHANGELOG.rst
@@ -15,9 +15,8 @@ Changed
 - validate training data only if used for training
 - applied spacy guidelines on how to disable pipeline components
 - starter packs now also tested when attempting to merge a branch to master
-
-=======
 - replace pep8 with pycodestyle
+- removed ``--num_threads`` from run command
 
 Removed
 -------